// src/app/services/delivery-manager/model/order.ts
/**
 * API Delivery Manager
 * API Delivery Manager
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Customer } from './customer';
import { Item } from './item';
import { Operation } from './operation';
import { Rider } from './rider';
import { Status } from './status';

/**
 * Order information
 */
export interface Order {
  /**
   * Order id
   */
  id?: string;
  /**
   * External Order Id
   */
  externalId?: string;
  /**
   * Order creation date
   */
  date?: string;
  /**
   * Indicates if it is an scheduled order
   */
  scheduled?: boolean;
  /**
   * Order type
   */
  type?: Order.TypeEnum;
  /**
   * Order channel
   */
  channel?: Order.ChannelEnum;
  /**
   * Order amount in cents without decimals
   */
  amount?: number;
  /**
   * Order notes
   */
  notes?: string;
  rider?: Rider;
  customer?: Customer;
  operation?: Operation;
  items?: Array<Item>;
  status?: Status;
  statuses?: Array<Status>;
}
export namespace Order {
  export type TypeEnum = 'Delivery' | 'Pickup' | 'Dinein';
  export const TypeEnum = {
    Delivery: 'Delivery' as TypeEnum,
    Pickup: 'Pickup' as TypeEnum,
    Dinein: 'Dinein' as TypeEnum,
  };
  export type ChannelEnum = 'Web' | 'Glovo' | 'JustEat' | 'Uber' | 'Waitry';
  export const ChannelEnum = {
    Web: 'Web' as ChannelEnum,
    Glovo: 'Glovo' as ChannelEnum,
    JustEat: 'JustEat' as ChannelEnum,
    Uber: 'Uber' as ChannelEnum,
    Waitry: 'Waitry' as ChannelEnum,
  };
}